requirements.txt
azure-core
azure-functions
azure-identity
azure-servicebus
azure-storage-blob
blinker
certifi
cffi
charset-normalizer
click
colorama
cryptography
fire
Flask
fonttools
google-cloud-pubsub
gunicorn
idna
importlib_metadata
isodate
itsdangerous
Jinja2
lxml
MarkupSafe
msal
msal-extensions
numpy
opencv-python-headless
packaging
pdf2docx
portalocker
pycparser
PyJWT
pywin32;sys_platform == 'win32'
requests
six
termcolor
typing_extensions
urllib3
Werkzeug
zipp
